+++ cli/azd/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.3.0-beta.6 (Unreleased)
 
+- [[#1060]](https://github.com/Azure/azure-dev/pull/1060) Fix color rendering on Windows.
+
 ## 0.3.0-beta.5 (2022-10-26)
 
 ### Bugs Fixed